#3 Added support for FEN notation.

+++ Sue/Sue.Common/Model/Chessboard/IChessboard.cs
@@ -11,8 +11,8 @@ namespace Sue.Common.Model.Chessboard
         bool BlackKingsideCastlingAvailable { get; }
         bool BlackQueensideCastlingAvailable { get; }
         IChessboardField EnPassantTargetField { get; }
-        ushort HalfmoveClock { get; }
-        ushort FullmoveNumber { get; }
+        int HalfmoveClock { get; }
+        int FullmoveNumber { get; }
         IChessboardField GetChessboardField(File file, Rank rank);
         IChessPiece GetChessPiece(File file, Rank rank);
         IEnumerable<IChessPiece> GetChessPieces(Color color);
